requirements.txt
attrs
binapy
certifi
cffi
charset-normalizer==3.1.0
cryptography
ecdsa
furl
idna
jwskate
orderedmultidict
pyasn1
pycparser
python-jose
requests==2.29.0
requests-oauth2client
retry
rsa
six
typing-extensions
urllib3==1.26.18